requirements.txt
streamlit==1.42.2
pandas==2.2.3
numpy==2.2.3
google-genai==1.2.0
pydantic==2.10.6
sift-stack-py
